requirements.txt
geopy==2.4.1
gpxpy==1.6.2
matplotlib==3.9.2
numpy==2.1.0
Pillow==9.5.0
Pillow==10.4.0
